+++ cate/ds/esa_cci_odp.py
@@ -56,7 +56,7 @@ from owslib.namespaces import Namespaces
 from cate.conf import get_config_value, get_data_stores_path
 from cate.conf.defaults import NETCDF_COMPRESSION_LEVEL
 from cate.core.ds import DATA_STORE_REGISTRY, DataAccessError, DataStore, DataSource, Schema, open_xarray_dataset
-from cate.core.opimpl import subset_spatial_impl, normalize_impl
+from cate.core.opimpl import subset_spatial_impl, normalize_impl, adjust_spatial_attrs_impl
 from cate.core.types import PolygonLike, TimeLike, TimeRange, TimeRangeLike, VarNamesLike
 from cate.ds.local import add_to_data_store_registry, LocalDataSource, LocalDataStore
 from cate.util.monitor import Cancellation, Monitor
@@ -636,7 +636,6 @@ class EsaCciOdpDataSource(DataSource):
                      var_names: VarNamesLike.TYPE = None,
                      protocol: str = None) -> Any:
         time_range = TimeRangeLike.convert(time_range) if time_range else None
-        region = PolygonLike.convert(region) if region else None
         var_names = VarNamesLike.convert(var_names) if var_names else None
 
         selected_file_list = self._find_files(time_range)
@@ -692,7 +691,6 @@ class EsaCciOdpDataSource(DataSource):
 
         local_id = local_ds.id
         time_range = TimeRangeLike.convert(time_range)
-        region = PolygonLike.convert(region)
         var_names = VarNamesLike.convert(var_names)
 
         excluded_variables = get_exclude_variables_fix_known_issues(self.id)
@@ -745,24 +743,19 @@ class EsaCciOdpDataSource(DataSource):
                     remote_dataset = xr.open_dataset(dataset_uri, drop_variables=[variable.get('name') for variable in
                                                                                   excluded_variables])
                     if var_names:
-                        remote_dataset = remote_dataset.drop(
-                            [var_name for var_name in remote_dataset.data_vars.keys()
-                             if var_name not in var_names])
+                        remote_dataset = remote_dataset.drop([var_name for var_name in remote_dataset.data_vars.keys()
+                                                              if var_name not in var_names])
 
                     if region:
                         remote_dataset = normalize_impl(remote_dataset)
-                        remote_dataset = subset_spatial_impl(remote_dataset, region)
-                        geo_lon_min, geo_lat_min, geo_lon_max, geo_lat_max = region.bounds
+                        remote_dataset = adjust_spatial_attrs_impl(subset_spatial_impl(remote_dataset, region),
+                                                                   allow_point=False)
 
-                        remote_dataset.attrs['geospatial_lat_min'] = geo_lat_min
-                        remote_dataset.attrs['geospatial_lat_max'] = geo_lat_max
-                        remote_dataset.attrs['geospatial_lon_min'] = geo_lon_min
-                        remote_dataset.attrs['geospatial_lon_max'] = geo_lon_max
                         if do_update_of_region_meta_info_once:
-                            local_ds.meta_info['bbox_maxx'] = geo_lon_max
-                            local_ds.meta_info['bbox_minx'] = geo_lon_min
-                            local_ds.meta_info['bbox_maxy'] = geo_lat_max
-                            local_ds.meta_info['bbox_miny'] = geo_lat_min
+                            local_ds.meta_info['bbox_minx'] = remote_dataset.attrs['geospatial_lon_min']
+                            local_ds.meta_info['bbox_maxx'] = remote_dataset.attrs['geospatial_lon_max']
+                            local_ds.meta_info['bbox_maxy'] = remote_dataset.attrs['geospatial_lat_max']
+                            local_ds.meta_info['bbox_miny'] = remote_dataset.attrs['geospatial_lat_min']
                             do_update_of_region_meta_info_once = False
 
                     if compression_enabled: